refactor: move permalink and full table mappers into separate files
