--- src/explore-education-statistics-common/src/modules/table-tool/utils/mapPermalinks.ts
@@ -1,133 +1,0 @@
-import { FilterOption } from '@common/modules/table-tool/services/tableBuilderService';
-import { transformTableMetaFiltersToCategoryFilters } from '@common/modules/table-tool/components/utils/tableToolHelpers';
-import {
-  Permalink,
-  UnmappedFullTable,
-  UnmappedPermalink,
-  UnmappedTableHeadersConfig,
-} from '@common/modules/table-tool/services/permalinkService';
-import {
-  CategoryFilter,
-  Indicator,
-  LocationFilter,
-  TimePeriodFilter,
-} from '@common/modules/table-tool/types/filters';
-import { FullTable, FullTableMeta } from '@common/modules/table-tool/types/fullTable';
-import { TableHeadersConfig } from '@common/modules/table-tool/utils/tableHeaders';
-
-const reverseMapTableHeadersConfig = (
-  { columns, rows, columnGroups, rowGroups }: UnmappedTableHeadersConfig,
-  fullTableSubjectMeta: FullTableMeta,
-): TableHeadersConfig => {
-  /**
-   * config filters only contain `values`.
-   * This function remaps the config filters into full filters, using the subjectMeta
-   */
-  let mappedRows: (TimePeriodFilter | Indicator)[] = [];
-  let mappedColumns: (TimePeriodFilter | Indicator)[] = [];
-
-  const initialValue = columns[0].value;
-  // rows/columns can only be TimePeriods / Indicators
-  if (Number.isNaN(Number(initialValue))) {
-    // if NaN then timePeriod
-    mappedColumns = columns.map(({ value }) => {
-      return fullTableSubjectMeta.timePeriodRange.find(
-        timePeriod => `${timePeriod.year}_${timePeriod.code}` === value,
-      ) as TimePeriodFilter;
-    });
-    mappedRows = rows.map(({ value }) => {
-      return fullTableSubjectMeta.indicators.find(
-        indicator => indicator.value === value,
-      ) as Indicator;
-    });
-  } else {
-    mappedRows = rows.map(({ value }) => {
-      return fullTableSubjectMeta.timePeriodRange.find(
-        timePeriod => `${timePeriod.year}_${timePeriod.code}` === value,
-      ) as TimePeriodFilter;
-    });
-    mappedColumns = columns.map(({ value }) => {
-      return fullTableSubjectMeta.indicators.find(
-        indicator => indicator.value === value,
-      ) as Indicator;
-    });
-  }
-
-  // rowGroups/columnGroups can only be filters and locations
-  const locationAndFilterGroups: (LocationFilter | CategoryFilter)[][] = [
-    ...Object.values(
-      transformTableMetaFiltersToCategoryFilters(fullTableSubjectMeta.filters),
-    ),
-    fullTableSubjectMeta.locations,
-  ];
-
-  const mapOptionGroupsToFilterGroups = (
-    optionsGroups: FilterOption[][],
-  ): (LocationFilter | CategoryFilter)[][] =>
-    optionsGroups.map(optionGroup => {
-      const currentIndex = locationAndFilterGroups.findIndex(options =>
-        options.find(element => element.value === optionGroup[0].value),
-      );
-
-      return optionGroup.map(
-        ({ value }) =>
-          locationAndFilterGroups[currentIndex].find(
-            element => element.value === value,
-          ) as LocationFilter | CategoryFilter,
-      );
-    });
-
-  return {
-    columns: mappedColumns,
-    columnGroups: mapOptionGroupsToFilterGroups(columnGroups),
-    rows: mappedRows,
-    rowGroups: mapOptionGroupsToFilterGroups(rowGroups),
-  };
-};
-
-export const mapFullTable = (
-  unmappedFullTable: UnmappedFullTable,
-): FullTable => {
-  const subjectMeta = unmappedFullTable.subjectMeta || {
-    indicators: [],
-    locations: [],
-    timePeriodRange: [],
-  };
-
-  return {
-    ...unmappedFullTable,
-    subjectMeta: {
-      filters: {},
-      ...unmappedFullTable.subjectMeta,
-      indicators: subjectMeta.indicators.map(
-        indicator => new Indicator(indicator),
-      ),
-      locations: subjectMeta.locations.map(
-        location => new LocationFilter(location, location.level),
-      ),
-      timePeriodRange: subjectMeta.timePeriodRange.map(
-        timePeriod => new TimePeriodFilter(timePeriod),
-      ),
-    },
-  };
-};
-
-export const mapPermalink = (
-  unmappedPermalink: UnmappedPermalink,
-): Permalink => {
-  const mappedFullTable = mapFullTable(unmappedPermalink.fullTable);
-
-  return {
-    ...unmappedPermalink,
-    fullTable: mappedFullTable,
-    query: {
-      ...unmappedPermalink.query,
-      configuration: {
-        tableHeadersConfig: reverseMapTableHeadersConfig(
-          unmappedPermalink.query.configuration.tableHeadersConfig,
-          mappedFullTable.subjectMeta,
-        ),
-      },
-    },
-  };
-};
