+++ src/explore-education-statistics-common/src/modules/table-tool/utils/mapTableHeadersConfig.ts
